% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings.R
\name{cfb_rankings_drew_elo}
\alias{cfb_rankings_drew_elo}
\title{Generate rankings based on performance scores}
\usage{
cfb_rankings_drew_elo(year = 2022, years_lookback = 5)
}
\arguments{
\item{year}{A given CFB season}

\item{years_lookback}{The number of previous years to include in elo scores}
}
\value{
A data frame with team names and their rankings.
}
\description{
This function generates rankings for college football teams based on their
performance scores. Teams with higher performance scores are ranked higher.
}
