% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings.R
\name{cfb_rankings_elo_expected_win_prob}
\alias{cfb_rankings_elo_expected_win_prob}
\title{Calculates an expected score given two teams' ratings}
\usage{
cfb_rankings_elo_expected_win_prob(team_rating, opp_team_rating)
}
\arguments{
\item{home_rating}{The home team's elo rating}

\item{away_rating}{The away team's elo rating}
}
\value{
A data frame with team names and their rankings.
}
\description{
Calculates an expected score given two teams' ratings
}
