% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings.R
\name{cfb_rankings_elo_new_score}
\alias{cfb_rankings_elo_new_score}
\title{Calculates a new elo score given a game's outcome}
\usage{
cfb_rankings_elo_new_score(
  team_rating,
  expected_score,
  observed_score,
  k_factor = 40
)
}
\arguments{
\item{team_rating}{The team's elo rating}

\item{expected_score}{The expected score of the game}

\item{observed_score}{The actual score of the game}

\item{k_factor}{The k-factor to be used}
}
\value{
A data frame with team names and their rankings.
}
\description{
Calculates a new elo score given a game's outcome
}
