% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_cleaning.R
\name{clean_rcfb_comments}
\alias{clean_rcfb_comments}
\title{Clean r/CFB comments}
\usage{
clean_rcfb_comments(data)
}
\arguments{
\item{data}{some r/CFB comment data}
}
\value{
a cleaned dataframe
}
\description{
Clean r/CFB comments
}
