% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_single_game_report.R
\name{generate_report_nfl}
\alias{generate_report_nfl}
\title{Generate a report based on a single game thread from r/NFL ==================}
\usage{
generate_report_nfl(thread_data, alt_color = FALSE)
}
\arguments{
\item{thread_data}{a locally saved .RDS file of reddit comment data}

\item{alt_color}{whether or not the home team should use their alt color}

\item{output_file}{the file to save the resulting image}

\item{year}{the nfl season during which the game took place}
}
\value{
A dataframe of the reddit comments
}
\description{
Generate a report based on a single game thread from r/NFL ==================
}
