% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_reddit_url.R
\name{scrape_reddit_url}
\alias{scrape_reddit_url}
\title{Scrape the comments from a reddit thread by URL}
\usage{
scrape_reddit_url(thread_url, save_locally = FALSE, save_local_directory)
}
\arguments{
\item{thread_url}{The URL you want to scrape}

\item{save_locally}{TRUE or FALSE; whether or not to save an RDS of the results locally}

\item{save_local_directory}{a string; the directory where the RDS file should be saved}
}
\value{
A dataframe of the reddit comments
}
\description{
Scrape the comments from a reddit thread by URL
}
