% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankings.R
\name{cfb_rankings_drew_elo_top}
\alias{cfb_rankings_drew_elo_top}
\title{Generate Top 25 ELO rankings for a given week}
\usage{
cfb_rankings_drew_elo_top(
  year = 2022,
  years_lookback = 5,
  week = 1,
  n_results = 25
)
}
\arguments{
\item{year}{A given CFB season}

\item{years_lookback}{The number of previous years to include in elo scores}

\item{week}{A given CFB week}

\item{n_results}{The number of results to return}
}
\value{
A data frame with team names and their rankings.
}
\description{
This function generates rankings for college football teams based on their
performance scores. Teams with higher performance scores are ranked higher.
}
